requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
black==23.11.0
celery==5.3.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
crispy-bootstrap5==2025.6
cron-descriptor==1.4.5
cryptography==45.0.6
cycler==0.12.1
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.57.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-crispy-forms==2.4
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-rosetta==0.9.9
django-timescaledb==0.2.13
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et_xmlfile==2.0.0
flake8==6.1.0
fonttools==4.59.0
gunicorn==21.2.0
idna==3.10
iniconfig==2.1.0
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.4
matplotlib==3.10.5
mccabe==0.7.0
mlxtend==0.22.0
mypy_extensions==1.1.0
numpy==1.25.2
oauthlib==3.3.1
openpyxl==3.1.2
packaging==25.0
pandas==2.1.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
plotly==5.15.0
pluggy==1.6.0
polib==1.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.1
pytest-django==4.6.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0
redis==5.0.1
requests==2.32.4
requests-oauthlib==2.0.0
scikit-learn==1.3.0
scipy==1.11.0
six==1.17.0
sqlparse==0.5.3
tenacity==9.1.2
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
xlrd==2.0.1
XlsxWriter==3.1.0
